Address PR review: extract theme + add generation tokens
- Extract cmuxMarkdownTheme to Sources/Panels/MarkdownPanelTheme.swift
  to bring MarkdownPanelView.swift below the 400-line file budget
  (was 474, now 330).
- Add per-button generation tokens to copyFilePath and
  copyMarkdownContent so that a rapid second click does not let the
  first asyncAfter reset clobber the fresh checkmark, matching the
  triggerFocusFlashAnimation pattern in the same file.
