+++ Sources/Panels/MarkdownPanelTheme.swift
@@ -1,0 +1,145 @@
+import SwiftUI
+import MarkdownUI
+
+enum MarkdownPanelTheme {
+    static func make(colorScheme: ColorScheme) -> Theme {
+        let isDark = colorScheme == .dark
+
+        return Theme()
+            .text {
+                ForegroundColor(isDark ? .white.opacity(0.9) : .primary)
+                FontSize(14)
+            }
+            .heading1 { configuration in
+                VStack(alignment: .leading, spacing: 8) {
+                    configuration.label
+                        .markdownTextStyle {
+                            FontWeight(.bold)
+                            FontSize(28)
+                            ForegroundColor(isDark ? .white : .primary)
+                        }
+                    Divider()
+                }
+                .markdownMargin(top: 24, bottom: 16)
+            }
+            .heading2 { configuration in
+                VStack(alignment: .leading, spacing: 6) {
+                    configuration.label
+                        .markdownTextStyle {
+                            FontWeight(.bold)
+                            FontSize(22)
+                            ForegroundColor(isDark ? .white : .primary)
+                        }
+                    Divider()
+                }
+                .markdownMargin(top: 20, bottom: 12)
+            }
+            .heading3 { configuration in
+                configuration.label
+                    .markdownTextStyle {
+                        FontWeight(.semibold)
+                        FontSize(18)
+                        ForegroundColor(isDark ? .white : .primary)
+                    }
+                    .markdownMargin(top: 16, bottom: 8)
+            }
+            .heading4 { configuration in
+                configuration.label
+                    .markdownTextStyle {
+                        FontWeight(.semibold)
+                        FontSize(16)
+                        ForegroundColor(isDark ? .white : .primary)
+                    }
+                    .markdownMargin(top: 12, bottom: 6)
+            }
+            .heading5 { configuration in
+                configuration.label
+                    .markdownTextStyle {
+                        FontWeight(.medium)
+                        FontSize(14)
+                        ForegroundColor(isDark ? .white : .primary)
+                    }
+                    .markdownMargin(top: 10, bottom: 4)
+            }
+            .heading6 { configuration in
+                configuration.label
+                    .markdownTextStyle {
+                        FontWeight(.medium)
+                        FontSize(13)
+                        ForegroundColor(isDark ? .white.opacity(0.7) : .secondary)
+                    }
+                    .markdownMargin(top: 8, bottom: 4)
+            }
+            .codeBlock { configuration in
+                ScrollView(.horizontal, showsIndicators: true) {
+                    configuration.label
+                        .markdownTextStyle {
+                            FontFamilyVariant(.monospaced)
+                            FontSize(13)
+                            ForegroundColor(isDark ? Color(red: 0.9, green: 0.9, blue: 0.9) : Color(red: 0.2, green: 0.2, blue: 0.2))
+                        }
+                        .padding(12)
+                }
+                .background(isDark
+                    ? Color(nsColor: NSColor(white: 0.08, alpha: 1.0))
+                    : Color(nsColor: NSColor(white: 0.93, alpha: 1.0)))
+                .clipShape(RoundedRectangle(cornerRadius: 6))
+                .markdownMargin(top: 8, bottom: 8)
+            }
+            .code {
+                FontFamilyVariant(.monospaced)
+                FontSize(13)
+                ForegroundColor(isDark ? Color(red: 0.85, green: 0.6, blue: 0.95) : Color(red: 0.6, green: 0.2, blue: 0.7))
+                BackgroundColor(isDark
+                    ? Color(nsColor: NSColor(white: 0.18, alpha: 1.0))
+                    : Color(nsColor: NSColor(white: 0.92, alpha: 1.0)))
+            }
+            .blockquote { configuration in
+                HStack(spacing: 0) {
+                    RoundedRectangle(cornerRadius: 1.5)
+                        .fill(isDark ? Color.white.opacity(0.2) : Color.gray.opacity(0.4))
+                        .frame(width: 3)
+                    configuration.label
+                        .markdownTextStyle {
+                            ForegroundColor(isDark ? .white.opacity(0.6) : .secondary)
+                            FontSize(14)
+                        }
+                        .padding(.leading, 12)
+                }
+                .markdownMargin(top: 8, bottom: 8)
+            }
+            .link {
+                ForegroundColor(Color.accentColor)
+            }
+            .strong {
+                FontWeight(.semibold)
+            }
+            .table { configuration in
+                configuration.label
+                    .markdownTableBorderStyle(.init(color: isDark ? .white.opacity(0.15) : .gray.opacity(0.3)))
+                    .markdownTableBackgroundStyle(
+                        .alternatingRows(
+                            isDark
+                                ? Color(nsColor: NSColor(white: 0.14, alpha: 1.0))
+                                : Color(nsColor: NSColor(white: 0.96, alpha: 1.0)),
+                            isDark
+                                ? Color(nsColor: NSColor(white: 0.10, alpha: 1.0))
+                                : Color(nsColor: NSColor(white: 1.0, alpha: 1.0))
+                        )
+                    )
+                    .markdownMargin(top: 8, bottom: 8)
+            }
+            .thematicBreak {
+                Divider()
+                    .markdownMargin(top: 16, bottom: 16)
+            }
+            .listItem { configuration in
+                configuration.label
+                    .markdownMargin(top: 4, bottom: 4)
+            }
+            .paragraph { configuration in
+                configuration.label
+                    .markdownMargin(top: 4, bottom: 8)
+            }
+    }
+}
